Add method ImportElementsFrom on inverse bloom filter
ReadFrom and GobDecode methods allowed a bloom filter to be
exported and then recreated as part of a new array. The new
ImportElementsFrom method uses the filter.Add method to add
the elements written out by one filter to the array of another
filter.

This allows for use cases such as converting from one
filter size to another. For example, rolling a size 1,000
filter into a size 1,000,000 filter.