Add AddN function to CountMinSketch
Implements an AddN function that adds N occurrences of an item to the Count-Min Sketch.

The AddN function works similar to `Add` but takes a uint64 parameter for the count. This avoid paying the cost of hashing and the mod (`%`) operator for sketches with many occurrences (~100K in my case).